compile_tisean: build only the TISEAN subset hctsa needs; fix real warnings
Investigated TISEAN's compile warnings end to end. Full picture was much
bigger than initially scoped: 372 Fortran + 4 C warnings, and one fatal
Fortran compile error (cluster.f -- a bad format string that older
gfortran tolerated as an extension but modern versions reject outright).

Root fix: grepped every Operations/*.m for actual TISEAN binary
invocations and cross-checked against a full parse of TISEAN's own
Makefiles. The true dependency list is 11 binaries (c1, c2d, c2g, c2t,
d2, nstat_z, false_nearest, boxcount, lyap_r, lyap_spec, poincare) --
compile_tisean.m's own verification list was checking only 7, silently
missing boxcount/lyap_r/lyap_spec/poincare (so those could fail to build
without being caught). Rewrote compile_tisean.m to build only this
subset directly via targeted 'make' calls in source_c/source_f, copying
binaries straight from there rather than routing through TISEAN's
'do_install' target (which the previous commit already had to patch for
an unrelated bug) or a temp-prefix 'make install'. This alone makes the
cluster.f fatal error irrelevant, since that file is never compiled.

Added -std=legacy for the Fortran compiler (mirroring the existing
-std=gnu89 for C): silences ~300 warnings for old-style shared/labeled
DO-loop termination that gfortran still fully supports as a GNU
extension -- not suppressing anything at risk of being wrong, just
pedantry about outdated-but-supported syntax.

Fixed the small number of genuinely worthwhile issues in files that are
actually needed:
- source_c/d2.c: silence an implicit ULONG_MAX->double conversion
  warning with an explicit cast (same value either way, ~5e-20 relative
  precision loss inherent to any 64-bit-int-to-double conversion).
- source_c/boxcount.c: give next_dim() a real prototype instead of a
  bare K&R-style forward declaration.
- source_f/c1.f, c2d.f, c2g.f, c2t.f, help.f, readfile.f: converted 6
  Hollerith constants (e.g. 4h#m= ) in format strings to standard quoted
  strings -- mechanical, zero-semantic-risk (both forms emit identical
  characters), verified via the actual d2->c2g/c2t pipeline that the
  emitted text is byte-identical, which matters because
  NL_TISEAN_d2.m's SUB_readTISEANout parses that exact '#m=' text.

Verified: all 11 binaries build with zero warnings/errors; the 5 hctsa
Operations that depend on them (NL_TISEAN_d2, NL_BoxCorrDim,
NL_LargestLyap, NL_LyapSpec, NL_PoincareSection) run and produce sane
numeric output; full compile_mex run passes 8/8 components.

+++ Toolboxes/Tisean_3.0.1/source_c/d2.c
@@ -157,7 +157,7 @@ void scramble(void)
   check_alloc(schelp=(double*)malloc(sizeof(double)*hlength));
 
   for (i=0;i<hlength;i++)
-    rz[i]=(double)(rnd=rnd*rndf+1)/ULONG_MAX;
+    rz[i]=(double)(rnd=rnd*rndf+1)/(double)ULONG_MAX;
   
   for (i=0;i<SCBOX;i++)
     scbox[i]= -1;
